DiagramProfile refactored
