+++ src/main/java/org/reactome/server/tools/diagram/exporter/pptx/model/Gene.java
@@ -2,7 +2,7 @@ package org.reactome.server.tools.diagram.exporter.pptx.model;
 
 import com.aspose.slides.*;
 import org.reactome.server.tools.diagram.data.layout.Node;
-import org.reactome.server.tools.diagram.data.profile.DiagramProfile;
+import org.reactome.server.tools.diagram.data.profile.diagram.DiagramProfile;
 
 import java.awt.*;
