+++ src/main/java/org/reactome/server/tools/diagram/exporter/pptx/model/PPTXReaction.java
@@ -4,7 +4,7 @@ import com.aspose.slides.*;
 import org.reactome.server.tools.diagram.data.layout.Connector;
 import org.reactome.server.tools.diagram.data.layout.*;
 import org.reactome.server.tools.diagram.data.layout.Shape;
-import org.reactome.server.tools.diagram.data.profile.DiagramProfile;
+import org.reactome.server.tools.diagram.data.profile.diagram.DiagramProfile;
 
 import java.awt.Color;
 import java.util.HashMap;
