+++ src/main/java/org/reactome/server/tools/diagram/exporter/pptx/parser/DiagramPresentation.java
@@ -2,7 +2,7 @@ package org.reactome.server.tools.diagram.exporter.pptx.parser;
 
 import com.aspose.slides.*;
 import org.reactome.server.tools.diagram.data.layout.*;
-import org.reactome.server.tools.diagram.data.profile.DiagramProfile;
+import org.reactome.server.tools.diagram.data.profile.diagram.DiagramProfile;
 import org.reactome.server.tools.diagram.exporter.DiagramExporter;
 import org.reactome.server.tools.diagram.exporter.pptx.model.*;
 import org.slf4j.Logger;
